openflow: Refactor and simplify the OpenFlow tables (Attempt 2)
The OpenFlow tables were needlessly complicated due to a slightly too
clever usage of registers.  This simplifies it without changing the
functionality in any notable way.  This patch also lays the groundwork
for a future patch in which the OpenFlow tables are augmented to
enforce public ACLs.

The first version of this patch (commit e10f2df48) had an error that
caused it to break all network communication and was reverted.  This
version fixed the error.