quiltctl: Make minion role a command argument
Before, the minion role was simply a flag, but the command would fail if no role
was specified. Now, the role can also be specified as an argument. This way, the
usage string can clearly reflect that the role is required.

The flag option is preserved for cases when --role is good for readibality.