tests: Ignore image pull log warnings
The quay.io registry that etcd is hosted on is pretty flaky which
causes us to log a warning when we fail to pull. Later the
system retries and successfully runs the tests, but quilt-tester fails
due to the log message.  This log message isn't due to a bug in
Quilt, and logging the error is, in fact, the correct behavior when an
image fails to pull.  Therefore, this patch ignores it.