docker: Don't log twice on image pull error
When there's an error pulling an image, we log once in
minion/docker/docker.go, and once in its caller:
minion/scheduler/worker.go.  This patch changes it so the error
returned by minion/docker/docker makes clear that it's an image pull
error, and allows the caller to do the actual logging.