cloud: Abandon fork-join parallelism for thread-per-region
Before this patch the cloud provider code was mostly single-threaded,
but would fan out multiple threads when interacting with the
underlying provider.  This approach had two problems: First, the
performance was limited by the slowest region in the slowest cloud
provider.  Second, this approach added a bunch of unnecessary
complexity.

Instead, this patch makes a thread per region, both simplifying the
code, and making it feel snappier.