cloud: No longer set up acls between each machine pair
In a half hearted attempt to allow inter-region connectivity, we had
set up ACLs allowing explicit communication with every machine.  These
ACLs are unnecessary if every machine is in the same region (which is
the only configuration that Quilt supports at the moment).
Furthermore, they slightly complicated the code, and they had scaling
issues (due to a limit on the maximum number of ACLs supported by the
cloud providers).  This patch removes the functionality.