cloud: Set "Booting" status in the main join() transaction
Before this patch, the machine booting status was set outside of the
main join transaction.  This was not technically safe as the namespace
could have changed before that second transaction which could corrupt
the database.   Furthermore, the code was unnecessarily complicated,
and won't work with some upcoming cloud package changes.