cloud: Make provider functions non-blocking
Before this patch, each provider was expected to block until requests
made to it had completed and were visible in the List() function.
This required them to implement a lot of duplicate functionality that
was a bit finicky to get right. This patch now allows the cloud
providers to return immediately, and causes the cloud package to do
the waiting instead.